Add extra navigation drawer
The extra navigation drawer at the right side of a page. Add two new
views: One for managing products, one for managing categories.
Initially, the views only hold the two dialogs for adding products and
categories. Move the app bar buttons for opening the two dialogs to
the new views. Add the views to the router index.